requirements.txt
setuptools>=40.8.0
streamlit
fastapi==0.111.0
uvicorn==0.29.0
langchain
PyPDF2
langchain-google-genai
langchain_community
google-generativeai
python-dotenv
pydantic
faiss-cpu